Fix initial scroll for some browsers (#7059)
For some browsers (I tested desktop Edge and mobile Safari), setting document.scrollingElement.scrollTop to zero does not seem to be enough to reset the scroll position at start. Setting history.scrollRestoration = 'manual' seems to fix it for these browsers.
Firefox seems to work without this fix but works also with it.